+++ Codigos/algorithms.hpp
@@ -4,16 +4,16 @@
 using namespace std;
 
 
-std::vector<int> BubbleSortOptimized(vector<int> &vet, int n);
+std::vector<long long> BubbleSortOptimized(vector<int> &vet, int n);
 
 
-std::vector<int> BubbleSort(vector<int> &vet, int n);
+std::vector<long long> BubbleSort(vector<int> &vet, int n);
 
 
-std::vector<int> InsertionSort(vector<int> &vet, int n);
+std::vector<long long> InsertionSort(vector<int> &vet, int n);
 
 
-std::vector<int> InsertionSortOptimized(vector<int> &vet, int n);
+std::vector<long long> InsertionSortOptimized(vector<int> &vet, int n);
 
 
-std::vector<int> SelectionSort(vector<int> &vet, int n);
+std::vector<long long> SelectionSort(vector<int> &vet, int n);
