+++ Codigos/searches.hpp
@@ -2,6 +2,6 @@
 
 using namespace std;
 
-std::vector<int> binarySearch(int elem, vector<int> &vet, int n);
+std::vector<long long> binarySearch(int elem, vector<int> &vet, int n);
 
-std::vector<int> linearSearch(int elem, vector<int> &vet, int n);
+std::vector<long long> linearSearch(int elem, vector<int> &vet, int n);
