+++ Codigos/timeMeasure.hpp
@@ -4,6 +4,6 @@
 
 using namespace std;
 
-void timeMean(std::vector<int> (*func)(std::vector<int>&, int), std::vector<int> vet, int n, const std::string& nome);
+void timeMean(std::vector<long long> (*func)(std::vector<int>&, int), std::vector<int> vet, int n, const std::string& nome);
 
-void timeSearch(std::vector<int> (*searchFunc)(int, std::vector<int>&, int), int elem, std::vector<int> vet, int n, const std::string& nome);
+void timeSearch(std::vector<long long> (*searchFunc)(int, std::vector<int>&, int), int elem, std::vector<int> vet, int n, const std::string& nome);
